Add sitemap
This builds a proper `sitemap.xml` file at the root of the website. It
aids in SEO, telling search engines about the organization of the site's
content.

https://support.google.com/webmasters/answer/156184

The `URL` environment variable gives us an easy global way of accessing
the site's URL, for use in the sitemap and anywhere else. Netlify
provides some pre-defined variables, including `URL`, which "represents
the main address to your site". Using this means we won't need to set
the variable manually in the Netlify settings.

https://www.netlify.com/docs/continuous-deployment/#build-environment-variables